+++ app/src/main/java/com/akerimtay/smartwardrobe/weather/data/api/model/LocationResponse.kt
@@ -3,6 +3,6 @@ package com.akerimtay.smartwardrobe.weather.data.api.model
 import com.google.gson.annotations.SerializedName
 
 data class LocationResponse(
-    @SerializedName("lon") val longitude: Double,
-    @SerializedName("lat") val latitude: Double
+    @SerializedName("lon") val longitude: Float,
+    @SerializedName("lat") val latitude: Float
 )
